:recycle: created AutoPydanticHandler so AutoIpywidget can handle pydantic models
